specify XDG_CONFIG_HOME in docker file
This change is to improve the compatibility of the qiime 2 docker
image when used with the CWL reference runner(cwltool):
https://github.com/common-workflow-language/cwltool

This environment variable is used by the click module to determine
an app directory. Qiime2 uses this app directory to create
the qiime2 cache directory. When the qiime2 docker image is built it
populates the cache: `qiime dev refresh-cache`. This change will
allow use of this cached data by cwltool.

The CWL reference runner tries to run docker in a 'clean' enviroment.
Part of this includes running as user 502:20.
One simple change you can see is the following command no longer fails:
docker run --user 502:20 -it qiime2/core:2018.6 qiime info

See https://forum.qiime2.org/t/qiime-docker-image-always-caching-when-used-with-cwltool/4919